requirements.txt
tensorflow==3.5.6
matplotlib
Pillow
pywavelets
numpy==1.13.0